API Reference

Base

IMASDD.info — Function

info(uloc::String, extras::Bool=true)

Return information of a node in the IMAS data structure, possibly including extra structures

info(@nospecialize(ids::Union{IDS,IDSvector}), field::Symbol)::Dict{String,Any}

Return information of a filed of an IDS

IMASDD.units — Function

units(uloc::String)::String

Return string with units for a given IDS location

units(ids::IDS, field::Symbol)::String

Return string with units for a given IDS field

IMASDD.coordinates — Function

coordinates(@nospecialize(ids::IDS), field::Symbol; coord_leaves::Union{Nothing,Vector{Symbol}}=nothing)

Return two lists, one of coordinate names and the other with their values in the data structure

Coordinate value is nothing when the data does not have a coordinate

Coordinate value is missing if the coordinate is missing in the data structure

Use coord_leaves to override fetching coordinates of a given field

IMASDD.access_log — Constant

IMASDD.access_log

IMASDD.access_log.enable = true / false

@show IMASDD.access_log

empty!(IMASDD.access_log) # to reset

Track access to the data dictionary

Base.getproperty — Function

getproperty(@nospecialize(ids::IDS), field::Symbol)

Return IDS value for requested field

getproperty(@nospecialize(ids::IDS), field::Symbol, @nospecialize(default::Any))

Return IDS value for requested field or default if field is missing

NOTE: This is useful because accessing a missing field in an IDS would raise an error

Base.isempty — Function

isempty(@nospecialize(ids::IDSvector))::Bool

returns true if IDSvector is empty

isempty(@nospecialize(ids::IDS); include_expr::Bool=false, eval_expr::Bool=false)::Bool

Returns true if none of the IDS fields downstream have data (or expressions)

NOTE: By default it does not include nor evaluate expressions

isempty(@nospecialize(ids::IDS), field::Symbol; include_expr::Bool=false, eval_expr::Bool=false)::Bool

Returns true if the ids field has no data (or expression)

NOTE: By default it does not include nor evaluate expressions

Base.setproperty! — Function

Base.setproperty!(ids::IDS, field::Symbol, v; skip_non_coordinates::Bool=false, error_on_missing_coordinates::Bool=true)

Base.setproperty!(@nospecialize(ids::IDS), field::Symbol, v::AbstractArray{<:IDS}; skip_non_coordinates::Bool=false, error_on_missing_coordinates::Bool=true)

Handle setproperty of entire vectors of IDS structures at once (ids.field is of type IDSvector)

Base.setproperty!(
    @nospecialize(ids::IDS),
    field::Symbol,
    v::Union{AbstractRange,StaticArraysCore.SVector,StaticArraysCore.MVector};
    skip_non_coordinates::Bool=false,
    error_on_missing_coordinates::Bool=true
)

Convert abstract ranges and static arrays to vectors

Base.setproperty!(@nospecialize(ids::IDS), field::Symbol, v::AbstractArray; skip_non_coordinates::Bool=false, error_on_missing_coordinates::Bool=true)

Ensures coordinates are set before the data that depends on those coordinates. If skip_non_coordinates is set, then fields that are not coordinates will be silently skipped.

IMASDD.index — Function

index(@nospecialize(ids::IDSvectorElement))

Returns index of the IDSvectorElement in the parent IDSvector

IMASDD.keys_no_missing — Function

keys_no_missing(@nospecialize(ids::IDS); include_expr::Bool=true, eval_expr::Bool=false)

Returns generator of fields with data in a IDS

NOTE: By default it includes expressions, but does not evaluate them

Base.resize! — Function

Base.resize!(@nospecialize(ids::IDSvector{T}), condition::Pair{String}, conditions::Pair{String}...; wipe::Bool=true, error_multiple_matches::Bool=true)::T where {T<:IDSvectorElement}

Resize if a set of conditions are not met.

If wipe=true and an entry matching the condition is found, then the content of the matching IDS is emptied.

Either way, the IDS is populated with the conditions.

NOTE: error_multiple_matches will delete all extra entries matching the conditions.

Returns the selected IDS

Base.deleteat! — Function

Base.deleteat!(@nospecialize(ids::T), condition::Pair{String}, conditions::Pair{String}...)::T where {T<:IDSvector}

If an entry matching the condition is found, then the content of the matching IDS is emptied

Base.ismissing — Function

Base.ismissing(@nospecialize(ids::IDS), field::Symbol)::Bool

returns true/false if field is missing in IDS

Base.diff — Function

Base.diff(
    @nospecialize(ids1::T),
    @nospecialize(ids2::T);
    tol::Float64=1E-2,
    recursive::Bool=true,
    verbose::Bool=false) where {T<:IDS}

Compares two IDSs and returns dictionary with differences

NOTE: This function does not evaluate expressions (use freeze() on the IDSs to compare values instead of functions)

IMASDD.top_ids — Function

top_ids(@nospecialize(ids::Union{IDS,IDSvector}))::Union{<:IDS,Nothing}

Return top-level IDS in the hierarchy and nothing if top level is not a top-level IDS

IMASDD.top_dd — Function

top_dd(@nospecialize(ids::Union{IDS,IDSvector}))::Union{<:DD,Nothing}

Return top-level dd in the hierarchy, and nothing if top level is not dd

Base.parent — Function

parent(ids::Union{IDS,IDSvector}; IDS_is_absolute_top::Bool=true)

Return parent IDS/IDSvector in the hierarchy

If IDS_is_absolute_top=true then returns nothing instead of dd()

IMASDD.goto — Function

goto(@nospecialize(ids::Union{IDS,IDSvector}), loc::String)

Reach location in a given IDS

IMASDD.leaves — Function

leaves(@nospecialize(ids::IDS))

Returns iterator with (filled) leaves in the IDS

IMASDD.filled_ids_fields — Function

filled_ids_fields(@nospecialize(ids::IDS); eval_expr::Bool=false)::Vector{Tuple{<:IDS,Symbol}}

Returns a vector with tuples pointing to all the (ids, field) that have data downstream

IMASDD.paths — Function

paths(@nospecialize(ids::IDS); eval_expr::Bool=false)

Returns the locations in the IDS that have data downstream

IMASDD.selective_copy! — Function

selective_copy!(@nospecialize(h_in::IDS), @nospecialize(h_out::IDS), path::Vector{String}, time0::Float64)

Copies the content of a path from one IDS to another (if the path exists) at a given time0

NOTE:

the path is a i2p(ulocation)

if time0 is NaN then all times are retained

IMASDD.selective_delete! — Function

selective_delete!(@nospecialize(h_in::IDS), path::Vector{String})

Deletes a path from one IDS

NOTE:

the path is a i2p(ulocation)

IO

IMASDD.dict2imas — Function

dict2imas(dct::AbstractDict, @nospecialize(ids::IDS); verbose::Bool=false)

Populate IMAS data structure ids based on data contained in Julia dictionary dct.

Arguments

verbose::Bool=false: print structure hierarchy as it is filled

IMASDD.imas2dict — Function

imas2dict(ids::Union{IDS,IDSvector}; freeze::Bool=true, strict::Bool=false)

Populate Julia structure of dictionaries and vectors with data from IMAS data structure ids

IMASDD.json2imas — Function

json2imas(filename::AbstractString, @nospecialize(ids::IDS)=dd(); error_on_missing_coordinates::Bool=true, verbose::Bool=false)::IDS

Load the IMAS data structure from a JSON file with given filename

Arguments

error_on_missing_coordinates: boolean to raise an error if coordinate is missing

verbose: print structure hierarchy as it is filled

IMASDD.jstr2imas — Function

jstr2imas(json_string::String, @nospecialize(ids::IDS)=dd(); error_on_missing_coordinates::Bool=true, verbose::Bool=false)::IDS

Load the IMAS data structure from a JSON string

Arguments

error_on_missing_coordinates: boolean to raise an error if coordinate is missing

verbose: print structure hierarchy as it is filled

IMASDD.imas2json — Function

imas2json(@nospecialize(ids::Union{IDS,IDSvector}), filename::AbstractString; freeze::Bool=true, strict::Bool=false, indent::Int=0, kw...)

Save the IMAS data structure to a JSON file with given filename.

Arguments

freeze evaluates expressions

strict dumps fields that are strictly in ITER IMAS only

kw... arguments are passed to the JSON.print function

IMASDD.hdf2imas — Function

hdf2imas(filename::AbstractString; error_on_missing_coordinates::Bool=true, kw...)

Load data from a HDF5 file generated by OMAS Python platform (ie. hierarchical HDF5)

Arguments

kw... arguments are passed to the HDF5.h5open function

IMASDD.imas2hdf — Function

imas2hdf(@nospecialize(ids::Union{IDS,IDSvector}), filename::AbstractString; freeze::Bool=true, strict::Bool=false, kw...)

Save the IMAS data structure to a OMAS HDF5 file with given filename (ie. hierarchical HDF5)

Arguments

kw... arguments are passed to the HDF5.h5open function

freeze evaluates expressions

strict dumps fields that are strictly in ITER IMAS only

IMASDD.h5i2imas — Function

h5i2imas(filename::AbstractString, @nospecialize(ids::IDS)=dd(); kw...)::IDS

Load data from a HDF5 file generated by IMAS platform (ie. tensorized HDF5)

Arguments

kw... arguments are passed to the HDF5.h5open function

IMASDD.imas2h5i — Function

imas2h5i(@nospecialize(ids::Union{IDS,IDSvector}), filename::AbstractString;
    freeze::Bool=true, strict::Bool=false, run::Int=0, shot::Int=0, hdf5_backend_version::String="1.0", kw...)

Save data to a HDF5 file generated by IMAS platform (ie. tensorized HDF5)

Arguments

kw... arguments are passed to the HDF5.h5open function

freeze evaluates expressions

strict dumps fields that are strictly in ITER IMAS only

run, shot, hdf5_backend_version arguments are used to set the HDF5 attributes

Time

IMASDD.global_time — Function

global_time(ids::Union{IDS,IDSvector})

Get the dd.global_time of a given IDS

If top-level dd cannot be reached then returns Inf

IMASDD.set_time_array — Function

set_time_array(@nospecialize(ids::IDS), field::Symbol, value)

Set value of a time-dependent array at the dd.global_time

set_time_array(@nospecialize(ids::IDS), field::Symbol, time0::Float64, value)

Set value of a time-dependent array at time0

IMASDD.get_time_array — Function

get_time_array(ids, field)

Get data from a time-dependent array at the dd.global_time

get_time_array(@nospecialize(ids::IDS), field::Symbol, time0::Float64, scheme::Symbol=:linear)

Get data from time dependent array

NOTE: logic for @ddtime array handling:

scheme (i) interpolation between array bounds

constant (c) extrapolation within bounds of time array

error (e) when time0 is before minimum(time)

For example:

time:   -oooo-
data:   -o-o--
ddtime: eiiicc

get_time_array(@nospecialize(ids::IDS), field::Symbol, time0::Vector{Float64}, scheme::Symbol=:linear)

IMASDD.@ddtime — Macro

@ddtime( X.Y )

Get data from time dependent array. Equivalent to:

get_time_array(X, :Y)

and

@ddtime( X.Y = V)

Set data in a time dependent array. Equivalent to:

set_time_array(X, :Y, V)

IMASDD.last_time — Function

last_time(dd::DD)::Float64

Returns the last time referenced in all the IDSs dd.XXX.time vectors (including dd.global_time)

IMASDD.last_global_time — Function

last_global_time(dd::DD)::Float64

Returns the last time referenced in all the IDSs dd.XXX.time vectors (including dd.global_time)

IMASDD.new_timeslice! — Function

new_timeslice!(ids::IDS, time0::Float64)

Recursively appends a deepcopy at time time0 of the last time-slice of all time-dependent array structures under a given ids

IMASDD.retime! — Function

retime!(ids::IDS, time0::Float64)

Recursively change the time of the last time-slices or last time-depedent vector elements in a IDS

Math

IMASDD.interp1d — Function

interp1d(x, y, scheme::Symbol=:linear)

One dimensional curve interpolations with sheme [:constant, :linear, :quadratic, :cubic, :pchip, :lagrange]

NOTE: this interpolation method will extrapolate

IMASDD.extrap1d — Function

extrap1d(itp::DataInterpolations.AbstractInterpolation; first=:extrapolate, last=:extrapolate) where {T<:Real}

first and last can be [:extrapolate, :flat, --value--] affect how the extrapolation is done at the either end of the array

IMASDD.gradient — Function

gradient(coord::AbstractVector{C}, arr::AbstractVector{A}; method::Symbol=:second_order) where {C<:Real, A<:Real}

The finite difference gradient. The returned gradient has the same shape as the input array.

method of the gradient can be one of [:thirdorder, :secondorder, :central, :backward, :forward]

For :central the gradient is computed using second order accurate central differences in the interior points and first order accurate one-sides (forward or backward) differences at the boundaries.

For :second_order the gradient is computed using second order accurate central differences in the interior points, and 2nd order differences at the boundaries.

For :third_order the gradient is computed from the cubic spline passing through the points

gradient(coord1::AbstractVector, coord2::AbstractVector, mat::Matrix, dim::Int; method::Symbol=:second_order)

Finite difference method of the gradient: [:second_order, :central, :backward, :forward]

Can be applied to either the first (dim=1) or second (dim=2) dimension

gradient(coord1::AbstractVector, coord2::AbstractVector, mat::Matrix)

Finite difference method of the gradient: [:second_order, :central, :backward, :forward]

Computes the gradient in both dimensions